% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkmeans.R
\name{gka_allele_mutation}
\alias{gka_allele_mutation}
\title{Allele mutation probability computation}
\usage{
gka_allele_mutation(data, k, centers, ...)
}
\arguments{
\item{data}{a set of observations, presented as a matrix-like object where
every row is a new observation. The matrix is of size \code{n} by \code{m}.}

\item{k}{the number of clusters.}

\item{centers}{a matrix of size \code{k} by \code{m} with the cluster centers}

\item{...}{additional arguments passed to \code{\link[proxy:dist]{proxy::dist()}}.}
}
\value{
a matrix of size \code{n} by \code{k} with the probability of each
allele mutating to a specific cluster.
}
\description{
Allele mutation probability computation
}
\author{
Eduardo Ruiz Sabajanes, \email{eduardo.ruizs@edu.uah.es}
}
