% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkmeans.R
\name{gka_centers}
\alias{gka_centers}
\title{Centroid computation}
\usage{
gka_centers(data, k, population)
}
\arguments{
\item{data}{a set of observations, presented as a matrix-like object where
every row is a new observation. The matrix is of size \code{n} by \code{m}.}

\item{k}{the number of clusters.}

\item{population}{a matrix of size \code{p} by \code{n} with the cluster
assignments for each observation.}
}
\value{
a 3D array of size \code{p} by \code{k} by \code{m}.
}
\description{
Centroid computation
}
\author{
Eduardo Ruiz Sabajanes, \email{eduardo.ruizs@edu.uah.es}
}
