% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkmeans.R
\name{gka_chromosome_fix}
\alias{gka_chromosome_fix}
\title{Chromosome fixing method}
\usage{
gka_chromosome_fix(population, k)
}
\arguments{
\item{population}{a matrix of size \code{p} by \code{n} with the cluster
assignments for each observation.}

\item{k}{the number of clusters.}
}
\value{
a matrix of size \code{p} by \code{n} with the cluster assignments
for each observation.
}
\description{
This method fixes chromosomes which do not have at least one
observation assigned to each cluster.
}
\author{
Eduardo Ruiz Sabajanes, \email{eduardo.ruizs@edu.uah.es}
}
