% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkmeans.R
\name{gka_fitness}
\alias{gka_fitness}
\title{Fitness function}
\usage{
gka_fitness(twcv)
}
\arguments{
\item{twcv}{a vector of size \code{p} with the total within cluster
variation of each individual in the population.}
}
\value{
a vector of size \code{p} with the fitness of each individual in the
population.
}
\description{
Fitness function
}
\author{
Eduardo Ruiz Sabajanes, \email{eduardo.ruizs@edu.uah.es}
}
