% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkmeans.R
\name{gka_initialization}
\alias{gka_initialization}
\title{Initialization method}
\usage{
gka_initialization(n, p, k)
}
\arguments{
\item{n}{the number of observations in the data.}

\item{p}{the number of individuals in the population.}

\item{k}{the number of clusters.}
}
\value{
a matrix of size \code{p} by \code{n} with the cluster assignments
for each observation.
}
\description{
Initialization method
}
\author{
Eduardo Ruiz Sabajanes, \email{eduardo.ruizs@edu.uah.es}
}
