% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkmeans.R
\name{gka_mutation}
\alias{gka_mutation}
\title{Mutation method}
\usage{
gka_mutation(chromosome, prob, k, mut_probability)
}
\arguments{
\item{chromosome}{a vector of size \code{n} with the cluster assignments for
each observation.}

\item{prob}{a matrix of size \code{n} by \code{k} with the probability of
each allele mutating to a specific cluster.}

\item{k}{the number of clusters.}

\item{mut_probability}{the probability of a mutation occurring.}
}
\value{
a vector of size \code{n} with the cluster assignments for each
observation i.e. a new chromosome.
}
\description{
Mutation method
}
\author{
Eduardo Ruiz Sabajanes, \email{eduardo.ruizs@edu.uah.es}
}
