% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkmeans.R
\name{gka_selection}
\alias{gka_selection}
\title{Selection method}
\usage{
gka_selection(p, fitness)
}
\arguments{
\item{p}{the number of individuals in the population.}

\item{fitness}{a vector of size \code{p} with the fitness of each individual
in the population.}
}
\value{
the index of the individual selected for reproduction.
}
\description{
Selection method
}
\author{
Eduardo Ruiz Sabajanes, \email{eduardo.ruizs@edu.uah.es}
}
